Added User Authentication with jwt

+++ graphql/resolvers/booking.js
@@ -2,7 +2,10 @@ const Booking = require('../../models/booking');
 const { transformEvent, transformBooking } = require('./merge');
 
 module.exports = {
-    bookings: async () => {
+    bookings: async (args, req) => {
+      if (!req.isAuth){
+        throw new Error('Unauthenticated!');
+      }
       try {
         const bookings = await Booking.find();
         return bookings.map(booking => {
@@ -12,27 +15,33 @@ module.exports = {
         throw err;
       }
     },
-    bookEvent: async args => {
+    bookEvent: async (args, req) => {
+      if (!req.isAuth){
+        throw new Error('Unauthenticated!');
+      }
         try {
-        const fetchedEvent = await Event.findOne({_id: args.eventId});
-        const booking = new Booking({
-            user: '5c3a0067dbb48f62181bea5d',
-            event: fetchedEvent
-        });
-        const result = await booking.save();
-        return transformBooking(result);
+          const fetchedEvent = await Event.findOne({_id: args.eventId});
+          const booking = new Booking({
+              user: req.userId,
+              event: fetchedEvent
+          });
+          const result = await booking.save();
+          return transformBooking(result);
         } catch (err) {
-        throw err;
+          throw err;
         }
     },
-    cancelBooking: async args => {
-        try {
+    cancelBooking: async (args, req) => {
+      if (!req.isAuth){
+        throw new Error('Unauthenticated!');
+      }
+      try {
         const booking = await Booking.findById(args.bookingId).populate('event');
         const event = transformEvent(booking.event);
         await Booking.deleteOne({ _id: args.bookingId });
         return event;
-        } catch (err) {
+      } catch (err) {
         throw err;
-        }
+      }
     }
 }
